### 2017-08-13 #353 @jordigh
Ohai. So, a long time I ago I stopped paying attention to JOSS notifications because it was too much of a firehose, and I recently unsubscribed. Getting individually approached for one revision at a time, however, seems much more manageable. This submission does look like something I can work on.

Sorry, I've been out of the loop. Do we have a rubric somewhere that I can use to evaluate this submission?
