### 2017-08-14 #353 @arfon
> This submission does look like something I can work on.

Great, thanks!

> Getting individually approached for one revision at a time, however, seems much more manageable.

Yes, we've improved our workflow a fair bit since you probably unsubscribed 😁

> Sorry, I've been out of the loop. Do we have a rubric somewhere that I can use to evaluate this submission?

Yes. I'll make the review issue now with the checklist and see you over there.
