### 2017-08-14 #365 @whedon
[REVIEW]: gmm_diag and gmm_full: C++ classes for multi-threaded Gaussian mixture models and Expectation-Maximisation
**Submitting author:** @conradsnicta (<a href="http://orcid.org/0000-0002-0049-4501">Conrad Sanderson</a>)
**Repository:** <a href="https://github.com/conradsnicta/armadillo-gmm" target ="_blank">https://github.com/conradsnicta/armadillo-gmm</a>
**Version:** v7.960
