### 2017-08-16 #365 @conradsnicta
@jordigh - I'm the first author. If you have any questions on this submission (code, documentation, etc), I'd be happy to answer them.
