### 2017-08-16 #365 @jordigh
Thanks, I've started looking it over, and at a glance it looks good. It's my first review, so I'm taking some time to go through each of the points. One question, the code is already in the main Armadillo repo and that's in Sourceforge, right? It wasn't immediately obvious to me where the canonical repo is.

Btw, I work on Octave, not too far afield from Armadillo. I'm a fan. :-)
