### 2017-08-16 #365 @conradsnicta
Yes, the gmm code is in the main repo.  We moved it from Sourceforge to GitHub a while back.  The unstable branch is https://github.com/conradsnicta/armadillo-code/tree/unstable which is intended for rapid evolution and experimental code.

Due to the rapid change in the unstable repo, I've setup a separate repo for the purposes of the review: https://github.com/conradsnicta/armadillo-gmm
This essentially contains a snapshot of the Armadillo 7.960.1 release, but the README file is changed to focus on the gmm_diag and gmm_full classes.  I hope this is okay.
