### 2017-09-20 #365 @jordigh
I've been busy with some other affairs. I should have time this weekend to complete this review.
