### 2017-09-22 #365 @conradsnicta
@jordigh - Thanks for the update.  I've added tests for the `gmm_diag` and `gmm_full` classes to the git repo: https://github.com/conradsnicta/armadillo-gmm.  The corresponding Zenodo archive of the release has also been updated: http://dx.doi.org/10.5281/zenodo.838397. (The DOI in the paper will automatically resolve to the latest version of the release archive).

The new tests are in `tests/gmm.cpp`  The instructions for compiling and running the tests are described in `tests/README.txt`.

The documentation for the `gmm_diag` and `gmm_full` classes is in the `docs.html` file within the repo and archive.  Please let me know if you have any questions on queries.
