### 2017-09-25 #365 @jordigh
So, forgive me for being a novice reviewer, but there's a couple of things that aren't clear to me:

* I'm not sure the paper states clearly who the target audience. It speaks broadly of how the problem could be applied to signal processing, statistics, and other fields. Is that sufficient? Any practitioner in those fields would be interested in this?

* At least a few DOIs seem to be missing in the references. For example, 10.1109/79.543975 for the Moon paper.

A few other remarks:

* I only tested the installation on Debian GNU/Linux, but it looks like a standard CMake build and the build dependencies are clearly stated for macOS and Windows. I'm confident it will work in other environments. The instructions seem correct. Only snag: I installed to a DESTDIR instead of systemwide, so I had to make sure that my linker was looking for `$DESTDIR/usr/lib/x86_64-linux-gnu` instead of just `$DESTDIR/usr/lib/`. I'm not sure if this is an odd peculiarity of my Debian system or a general point that should be mentioned in the installation instructions.

* I have managed to compile and execute the gmm example in the documentation. I would like to spend some time over the next few days understanding this example and the results I see to ensure it is reasonable.
