### 2017-09-26 #365 @conradsnicta
@jordigh - I've modified the paper as well as the installation instructions to address the issues you've raised. Details below:

> I'm not sure the paper states clearly who the target audience is. It speaks broadly of how the problem could be applied to signal processing, statistics, and other fields. Is that sufficient? Any practitioner in those fields would be interested in this?

Modelling of data via GMMs is a general statistical modelling approach, so it has a multitude of uses in many fields. Its target audience is hence any practitioner who needs to have a probabilistic model of multidimensional data and calculate likelihoods + etc using such a model.  From first hand experience I have used it for signal processing (eg. speech & images), pattern recognition and computer vision (eg. human action recognition). The paper mentions the use areas as examples, and these areas are not meant to be limiting. Nevertheless, I have tightened the first sentence in the paper to clearly state that this is a statistical modelling technique.

> At least a few DOIs seem to be missing in the references.
> For example, 10.1109/79.543975 for the Moon paper.

I've added the missing DOIs (the Moon paper as well as the Dean & Ghemawat paper).

> The instructions seem correct. Only snag: I installed to a DESTDIR instead of systemwide, so I had to make sure that my linker was looking for $DESTDIR/usr/lib/x86_64-linux-gnu instead of just $DESTDIR/usr/lib/. I'm not sure if this is an odd peculiarity of my Debian system or a general point that should be mentioned in the installation instructions.

I've modified the installation instructions to state that the 'lib' directory might be named differently on the user's system and provided examples: _On recent 64 bit Debian & Ubuntu systems it is 'lib/x86_64-linux-gnu'. On recent 64 bit Fedora & RHEL systems it is 'lib64'_.
