### 2017-10-12 #365 @jordigh
Okay, I've looked a bit more at the functionality of `gmm_diag`, and the results seem adequate. I ran the model a few times, looked at the resulting convex combination of multivariate normals, and they seemed to fit the data to a reasonable degree.

Ticking through the other tickboxes of my review, I couldn't find much of an indication of how to contribute to Armadillo except a pointer to the github repo, which I suppose might be enough of an indication in this day and age of near-universal git and pull requests.

I compiled and ran the tests for the two new functions, although I did kludge the Makefile to compile with `-I ../include` like the corresponding Makefile in the `examples/` directory so helpfully already does, plus change its linking option from `-larmadillo` to `-lblas llapacak`. I assume this is because the test suite is generally meant to be ran against a system-installed Armadillo unlike the examples. The two tests look reasonable and both passed. I understand there's no RNG seeding, as I had to explicitly add a call to `arma_rng::set_seed_random()` for my own examples, an addition of which might make it interesting to see if the tests still pass despite lack of complete determinism.

All in all, this looks complete and I can tick off the final box, but I do have one final question: you already have a citable paper for this software in the conference proceedings, so what was the motivation to submit it to JOSS in the first place?
