### 2017-10-12 #365 @conradsnicta
@jordigh - Thanks for doing the review. To answer your question: the previous conference paper only covers the `gmm_diag` class, while the JOSS article would cover both the `gmm_diag` class and the newer `gmm_full` class.  Furthermore, the previous paper is at a minor local conference on signal processing, while JOSS directly targets open source software (ie. it's a better fit), as well as having wider reach and prestige.

I will update the citation within the documentation to point to the JOSS article if it is accepted.
