### 2017-10-16 #365 @conradsnicta
@arfon - DOI of the archive: 10.5281/zenodo.1012627  (http://dx.doi.org/10.5281/zenodo.1012627)
Corresponding paper source: https://github.com/conradsnicta/armadillo-gmm/tree/master/paper/
